+++ src/api/integration/integration_test.go
@@ -2008,6 +2008,8 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 		return x
 	}
 
+	defaultChangeAddress := w.Entries[0].Address.String()
+
 	type testCase struct {
 		name                 string
 		req                  api.CreateTransactionRequest
@@ -2030,7 +2032,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2053,7 +2055,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2086,7 +2088,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2109,7 +2111,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2138,7 +2140,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2162,6 +2164,44 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 		},
 
 		{
+			// NOTE: this test will fail if "totalCoins - 1e3" does not require
+			// all of the outputs to be spent, e.g. if there is an output with
+			// "totalCoins - 1e3" coins in it.
+			// TODO -- Check that the wallet does not have an output of 0.001,
+			// because then this test cannot be performed, since there is no
+			// way to use all outputs and produce change in that case.
+			name: "valid request, manual one output with change, spend all, unspecified change address",
+			req: api.CreateTransactionRequest{
+				HoursSelection: api.HoursSelection{
+					Type: wallet.HoursSelectionTypeManual,
+				},
+				Wallet: api.CreateTransactionRequestWallet{
+					ID:       w.Filename(),
+					Password: password,
+				},
+				To: []api.Receiver{
+					{
+						Address: w.Entries[1].Address.String(),
+						Coins:   toDropletString(totalCoins - 1e3),
+						Hours:   "1",
+					},
+				},
+			},
+			outputs: []coin.TransactionOutput{
+				{
+					Address: w.Entries[1].Address,
+					Coins:   totalCoins - 1e3,
+					Hours:   1,
+				},
+				{
+					// Address omitted -- will be check later in the test body
+					Coins: 1e3,
+					Hours: remainingHours - 1,
+				},
+			},
+		},
+
+		{
 			name: "valid request, manual one output with change, don't spend all",
 			req: api.CreateTransactionRequest{
 				HoursSelection: api.HoursSelection{
@@ -2171,7 +2211,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2202,7 +2242,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2232,7 +2272,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2261,7 +2301,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2301,7 +2341,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Password: password,
 					UxOuts:   []string{unknownOutput.Hex()},
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2325,7 +2365,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Password: password,
 					UxOuts:   []string{nonWalletOutputs[0].Hash},
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2349,7 +2389,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Password: password,
 					UxOuts:   []string{walletOutputs[0].Hash},
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2374,7 +2414,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Password: password,
 					UxOuts:   []string{walletOutputs[0].Hash},
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2383,7 +2423,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					},
 				},
 			},
-			err:  "400 Bad Request - Insufficient coinhours for transaction outputs\n",
+			err:  "400 Bad Request - hours are not sufficient\n",
 			code: http.StatusBadRequest,
 		},
 
@@ -2403,7 +2443,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					// More complex cases should be covered by unit tests
 					UxOuts: walletOutputHashes,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2440,7 +2480,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					Password:  password,
 					Addresses: []string{testutil.MakeAddress().String()},
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2469,7 +2509,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					// More complex cases should be covered by unit tests
 					Addresses: addresses,
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[1].Address.String(),
@@ -2504,7 +2544,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password + "foo",
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2527,7 +2567,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: "",
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2551,7 +2591,7 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 					ID:       w.Filename(),
 					Password: password + "foo",
 				},
-				ChangeAddress: w.Entries[0].Address.String(),
+				ChangeAddress: &defaultChangeAddress,
 				To: []api.Receiver{
 					{
 						Address: w.Entries[0].Address.String(),
@@ -2586,7 +2626,27 @@ func TestLiveWalletCreateTransactionSpecific(t *testing.T) {
 			}
 
 			for i, o := range tc.outputs {
-				require.Equal(t, o.Address.String(), result.Transaction.Out[i].Address)
+				// The final change output may not have the address specified,
+				// if the ChangeAddress was not specified in the wallet params.
+				// Calculate it automatically based upon the transaction inputs
+				if o.Address.Null() {
+					require.Equal(t, i, len(tc.outputs)-1)
+					require.Nil(t, tc.req.ChangeAddress)
+
+					changeAddr := result.Transaction.Out[i].Address
+					// The changeAddr must be associated with one of the transaction inputs
+					changeAddrFound := false
+					for _, x := range result.Transaction.In {
+						if changeAddr == x.Address {
+							changeAddrFound = true
+							break
+						}
+					}
+
+					require.True(t, changeAddrFound)
+				} else {
+					require.Equal(t, o.Address.String(), result.Transaction.Out[i].Address)
+				}
 
 				coins, err := droplet.FromString(result.Transaction.Out[i].Coins)
 				require.NoError(t, err)
@@ -2777,7 +2837,7 @@ func TestLiveWalletCreateTransactionRandom(t *testing.T) {
 				Mode:        wallet.HoursSelectionModeShare,
 				ShareFactor: shareFactor,
 			},
-			ChangeAddress: changeAddress,
+			ChangeAddress: &changeAddress,
 			Wallet: api.CreateTransactionRequestWallet{
 				ID:       w.Filename(),
 				Password: password,
@@ -2799,7 +2859,7 @@ func TestLiveWalletCreateTransactionRandom(t *testing.T) {
 				Mode:        wallet.HoursSelectionModeShare,
 				ShareFactor: "0",
 			},
-			ChangeAddress: changeAddress,
+			ChangeAddress: &changeAddress,
 			Wallet: api.CreateTransactionRequestWallet{
 				ID:       w.Filename(),
 				Password: password,
@@ -2826,7 +2886,7 @@ func TestLiveWalletCreateTransactionRandom(t *testing.T) {
 				Mode:        wallet.HoursSelectionModeShare,
 				ShareFactor: "1",
 			},
-			ChangeAddress: changeAddress,
+			ChangeAddress: &changeAddress,
 			Wallet: api.CreateTransactionRequestWallet{
 				ID:       w.Filename(),
 				Password: password,
@@ -2851,7 +2911,7 @@ func TestLiveWalletCreateTransactionRandom(t *testing.T) {
 			HoursSelection: api.HoursSelection{
 				Type: wallet.HoursSelectionTypeManual,
 			},
-			ChangeAddress: changeAddress,
+			ChangeAddress: &changeAddress,
 			Wallet: api.CreateTransactionRequestWallet{
 				ID:       w.Filename(),
 				Password: password,
@@ -3528,6 +3588,8 @@ func TestDisableWalletApi(t *testing.T) {
 		return
 	}
 
+	changeAddress := testutil.MakeAddress().String()
+
 	type testCase struct {
 		name        string
 		method      string
@@ -3694,10 +3756,10 @@ func TestDisableWalletApi(t *testing.T) {
 					Wallet: api.CreateTransactionRequestWallet{
 						ID: "test.wlt",
 					},
-					ChangeAddress: testutil.MakeAddress().String(),
+					ChangeAddress: &changeAddress,
 					To: []api.Receiver{
 						{
-							Address: testutil.MakeAddress().String(),
+							Address: changeAddress,
 							Coins:   "0.001",
 							Hours:   "1",
 						},